+++ Cargo.toml
@@ -10,15 +10,22 @@ version = "0.1.0"
 coffee = {version = "0.4.1", features = ["vulkan"]}
 cpal = "0.13.5"
 docopt = "1.1.1"
+env_logger = "0.9"
+imgui = "0.8.2"
+imgui-wgpu = "0.19"
+imgui-winit-support = {version = "0.8", default-features = false, features = ["winit-26"]}
 lazy_static = "1.4.0"
+log = "0.4"
 maplit = "1.0.2"
+pixels = "0.9.0"
 serde = "1.0.137"
+winit = "0.26"
+winit_input_helper = "0.12"
 
 # Copied directly from Cargo.lock - is there a way to reference
 # sub-dependencies, or auto-resolve a version based on what's in Cargo.lock?
 image = "0.21.3"
 rand = "0.7.3"
-winit = "0.24.0"
 
 [dev-dependencies]
 pretty_assertions = "1.2.1"
